pythonbin/pyproject.toml
[tool.poetry]
name = "pythonbin"
version = "0.1.0"
description = ""
authors = ["Asim Ihsan"]
readme = "README.md"

[tool.poetry.scripts]
update-workspace = "pythonbin.git.update_workspace:main"
ghreview = "pythonbin.gh.ghreview:main"
compress-logs = "pythonbin.logs.compress:run_main"
load = "pythonbin.load.load:run_main"
updater = "pythonbin.updater.updater:run_main"
extract = "pythonbin.extract.extract:run_main"
transcript = "pythonbin.transcript.main:run_main"

[tool.poe.tasks.scrape]
script = "pythonbin.scraper.scrape:run_main"
help = "Scrape a URL"

[[tool.poe.tasks.scrape.args]]
name = "url"
help = "URL to scrape"
options = ["--url"]
type = "string"

[tool.poe.tasks.extract]
script = "pythonbin.extract.extract:run_main"
help = "Extract text from file"

[[tool.poe.tasks.extract.args]]
name = "path"
help = "Path to file"
type = "string"
positional = true

[[tool.poe.tasks.extract.args]]
name = "write"
help = "Write to TXT file"
options = ["--write"]
type = "boolean"

[tool.poetry.dependencies]
anthropic = "^0.25.7"
beautifulsoup4 = "^4.12.2"
boilerpy3 = "^1.0.7"
dateparser = "^1.2.0"
gitpython = "^3.1.43"
html5lib = "^1.1"
instructor = "^1.2.4"
jira = "^3.6.0"
jupyter = "^1.0.0"
librosa = "^0.10.1"
litellm = "^1.35.31"
markdown = "^3.6"
markdown-it-py = { extras = ["plugins"], version = "^3.0.0" }
marvin = "^2.3.3"
mdformat = "^0.7.17"
mistletoe = "^1.2.1"
numpy = "^1.26.4"
numpydoc = "^1.7.0"
opencv-python = "^4.9.0.80"
pathspec = "^0.12.1"
pillow = "^10.2.0"
playwright = "^1.43.0"
psutil = "^5.9.7"
pyautogui = "^0.9.54"
pydantic = "2.7"
pynput = "^1.7.6"
pyscreeze = "^0.1.30"
python = ">=3.11,<3.12"
questionary = "^2.0.1"
rich = "^13.7.0"
scipy = "^1.13.0"
semver = "^3.0.2"
tiktoken = "^0.6.0"
transformers = "^4.40.1"
tree-sitter = "^0.21.3"
unstructured = { extras = ["all-docs"], version = "^0.13.4" }
watchdog = "^4.0.0"
zstandard = "^0.22.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
pyright = "1.1.360"
pytest = "^8.2.0"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = [
    "pythonbin",
]
minversion = "6.0"
filterwarnings = [
    "error",
    "ignore::UserWarning",
]

[tool.black]
line-length = 120
target-version = ['py311', 'py312']
include = '\.pyi?$'  # Regex pattern to include files
exclude = '/(build|buck-out|dist|_build|\.eggs|\.git|\.hg|\.mypy_cache|\.nox|\.tox|\.venv|venv|\.svn|_build|buck-out|build|dist)/'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
